Revert "Merge branch 'feature/aws_uplaud' into feature/Client"
This reverts commit 270c1487a9e27739eff9dc7e200c5c17e1e46d6c, reversing
changes made to 506b1f29fc2b2706c3a491c30a78e37bfa16ab18.

+++ Backend/src/main/java/app/rest/UserController.java
@@ -35,14 +35,15 @@ public class UserController {
     }
 
 
-    @DeleteMapping("/users/{id}")
-    public ResponseEntity<User> deleteUser(@PathVariable Integer id, @RequestAttribute(value = JWTokenInfo.KEY) JWTokenInfo tokenInfo) {
+    @DeleteMapping("/users/{email}")
+    public ResponseEntity<User> deleteUser(@PathVariable String email, @RequestAttribute(value = JWTokenInfo.KEY) JWTokenInfo tokenInfo) {
 
-//        if(!tokenInfo.getUser()) {
-//            throw new AuthorizationException("only administrators can remove members");
-//        }
+        if(!tokenInfo.isAdmin()) {
+            throw new AuthorizationException("only administrators can remove members");
+        }
 
-        User user = userRepo.findById(id);
+        User user = getUserByEmail(email);
+
         userRepo.delete(user);
 
         return ResponseEntity.ok(user);
